+++ templates/Layout/Header/ProjectOptions.html.twig
@@ -14,7 +14,7 @@
         </span>
     </li>
 
-  {# TODO: Disable Report Project for now. Needs a separate flag in database - a new concept!
+
   {% if not my_project|default(false) %}
     <li id="top-app-bar__btn-report-project"
         class="btn top-app-bar-btn-options mdc-list-item mdc-ripple-upgraded"
@@ -25,7 +25,7 @@
       </span>
     </li>
   {% endif %}
-  #}
+
   </ul>
 
 {% endif %}
