+++ templates/Project/ProjectPage.html.twig
@@ -240,6 +240,23 @@
        data-trans-clipboard-success="{{ 'clipboard.success_project'|trans({}, 'catroweb') }}"
        data-trans-clipboard-fail="{{ 'clipboard.fail'|trans({}, 'catroweb') }}"
   ></div>
+  <div class="js-project-report"
+       data-project-id="{{ project.id }}"
+       data-path-report="{{ url('api_report_program', { id: project.id }) }}"
+       data-path-login="{{ url('login') }}"
+       data-trans-success="{{ 'report.success'|trans({}, 'catroweb') }}"
+       data-trans-error="{{ 'report.error'|trans({}, 'catroweb') }}"
+       data-trans-report="{{ 'report.report'|trans({}, 'catroweb') }}"
+       data-trans-cancel="{{ 'report.cancel'|trans({}, 'catroweb') }}"
+       data-trans-header="{{ 'report.header'|trans({}, 'catroweb') }}"
+       data-trans-reason="{{ 'report.reason'|trans({}, 'catroweb') }}"
+       data-trans-inappropriate="{{ 'report.inappropriate'|trans({}, 'catroweb') }}"
+       data-trans-copyright="{{ 'report.copyright'|trans({}, 'catroweb') }}"
+       data-trans-spam="{{ 'report.spam'|trans({}, 'catroweb') }}"
+       data-trans-dislike="{{ 'report.dislike'|trans({}, 'catroweb') }}"
+       data-const-ok="{{ "t" }}"
+       data-logged-in="{{ app.user ? true : false }}">
+  </div>
   <div class="js-project-description-credits"
        data-project-id="{{ project.id }}"
        data-has-description="{% if project.description %}true{% else %}false{% endif %}"
